Split CLI into multiple files for testing
